requirements.txt
aliyun-python-sdk-core==2.13.29
aliyun-python-sdk-geoip==1.0.3
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
idna==2.10
jmespath==0.10.0
numpy==1.19.3
pandas==1.1.4
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
urllib3==1.26.2
